Prep for mtg with AMN, AC

+++ analysis/load.wp.R
@@ -1,4 +1,4 @@
-load.wp <- function(e, g, fn.df='analysis/data/wallpapers-on-databrary.csv'){
+load.wp <- function(e, g, fn.df='analysis/data/wallpapers-on-databrary.csv', save=TRUE, img.dir='img/'){
   # Loads selected wallpaper image from Databrary
   
   require(dplyr)
@@ -10,9 +10,16 @@ load.wp <- function(e, g, fn.df='analysis/data/wallpapers-on-databrary.csv'){
   
   # Build URL and download
   url <- paste('https://nyu.databrary.org/slot', slot, '-/asset', asset, 'download?inline=true', sep='/')
-  z = tempfile()
-  download.file(url, z)
-  p <- readPNG(z)
-  file.remove(z)
+  if (save) {
+    z = paste(img.dir, paste(g, e, sep="-"), ".png", sep="")
+    download.file(url, z)
+    p <- readPNG(z)
+    sprintf('Saved to %s', z)
+  } else {
+    z = tempfile()
+    download.file(url, z)
+    p <- readPNG(z)
+    file.remove(z)
+  }
   p
 }
